backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
cloudinary
pillow
python-dotenv
feedparser
certifi
feedparser
cloudinary
psycopg2
